requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
install==1.3.5
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
